refactor(modal): colocate ui-kit modal into modal/ dir
modal.vue moved to modal/index.vue with modal-slot.vue and
modal-mode-config.ts colocated alongside it as slot.vue and
mode-config.ts. All non-template logic (scroll lock, esc handling,
transition hooks, recede/restore) extracted into a colocated
use-modal-stack.ts composable, leaving index.vue as template + wiring.
