+++ src/App.vue
@@ -4,7 +4,7 @@ import { useI18n } from 'vue-i18n'
 import { useNoticeStore } from '@/stores/notice-store'
 import NoticeToast from '@/components/ui-kit/notice/toast.vue'
 import NoticePanel from '@/components/ui-kit/notice/panel.vue'
-import UiModal from '@/components/ui-kit/modal.vue'
+import UiModal from '@/components/ui-kit/modal/index.vue'
 import { springScaleIn } from '@/utils/animations/modal'
 import { noticeToastListLeave } from '@/utils/animations/notice-toast'
 import audio_player from '@/sfx/player'
